refactor: Separate database integration from `trie` into its own crate
