+++ crates/trie/trie-db/src/hashed_cursor/mod.rs
@@ -1,0 +1,3 @@
+/// Default implementation of the hashed state cursor traits.
+mod default;
+pub use default::DatabaseHashedStorageCursor;
