+++ crates/trie/trie-db/src/prefix_set/mod.rs
@@ -1,0 +1,2 @@
+mod loader;
+pub use loader::PrefixSetLoader;
