+++ crates/trie/trie-db/src/updates.rs
@@ -1,0 +1,160 @@
+use derive_more::Deref;
+use reth_db::{
+    cursor::{DbCursorRO, DbCursorRW, DbDupCursorRO, DbDupCursorRW},
+    tables,
+    transaction::{DbTx, DbTxMut},
+};
+use reth_primitives::{
+    trie::{
+        BranchNodeCompact, HashBuilder, Nibbles, StorageTrieEntry, StoredBranchNode
+    },
+    B256,
+};
+use std::collections::{hash_map::IntoIter, HashMap, HashSet};
+use reth_trie::updates::{TrieKey, TrieOp};
+use reth_trie::walker::TrieWalker;
+
+/// The aggregation of trie updates.
+#[derive(Debug, Default, Clone, PartialEq, Eq, Deref)]
+pub struct TrieUpdates {
+    trie_operations: HashMap<TrieKey, TrieOp>,
+}
+
+impl<const N: usize> From<[(TrieKey, TrieOp); N]> for TrieUpdates {
+    fn from(value: [(TrieKey, TrieOp); N]) -> Self {
+        Self { trie_operations: HashMap::from(value) }
+    }
+}
+
+impl IntoIterator for TrieUpdates {
+    type Item = (TrieKey, TrieOp);
+    type IntoIter = IntoIter<TrieKey, TrieOp>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.trie_operations.into_iter()
+    }
+}
+
+impl TrieUpdates {
+    /// Schedule a delete operation on a trie key.
+    ///
+    /// # Panics
+    ///
+    /// If the key already exists and the operation is an update.
+    pub fn schedule_delete(&mut self, key: TrieKey) {
+        let existing = self.trie_operations.insert(key, TrieOp::Delete);
+        if let Some(op) = existing {
+            assert!(!op.is_update(), "Tried to delete a node that was already updated");
+        }
+    }
+
+    /// Extend the updates with trie updates.
+    pub fn extend(&mut self, updates: impl IntoIterator<Item = (TrieKey, TrieOp)>) {
+        self.trie_operations.extend(updates);
+    }
+
+    /// Extend the updates with account trie updates.
+    pub fn extend_with_account_updates(&mut self, updates: HashMap<Nibbles, BranchNodeCompact>) {
+        self.extend(
+            updates.into_iter().map(|(nibbles, node)| {
+                (TrieKey::AccountNode(nibbles.into()), TrieOp::Update(node))
+            }),
+        );
+    }
+
+    /// Finalize state trie updates.
+    pub fn finalize_state_updates<C>(
+        &mut self,
+        walker: TrieWalker<C>,
+        hash_builder: HashBuilder,
+        destroyed_accounts: HashSet<B256>,
+    ) {
+        // Add updates from trie walker.
+        let (_, walker_updates) = walker.split();
+        self.extend(walker_updates);
+
+        // Add account node updates from hash builder.
+        let (_, hash_builder_updates) = hash_builder.split();
+        self.extend_with_account_updates(hash_builder_updates);
+
+        // Add deleted storage tries for destroyed accounts.
+        self.extend(
+            destroyed_accounts.into_iter().map(|key| (TrieKey::StorageTrie(key), TrieOp::Delete)),
+        );
+    }
+
+    /// Finalize storage trie updates for a given address.
+    pub fn finalize_storage_updates<C>(
+        &mut self,
+        hashed_address: B256,
+        walker: TrieWalker<C>,
+        hash_builder: HashBuilder,
+    ) {
+        // Add updates from trie walker.
+        let (_, walker_updates) = walker.split();
+        self.extend(walker_updates);
+
+        // Add storage node updates from hash builder.
+        let (_, hash_builder_updates) = hash_builder.split();
+        self.extend(hash_builder_updates.into_iter().map(|(nibbles, node)| {
+            (TrieKey::StorageNode(hashed_address, nibbles.into()), TrieOp::Update(node))
+        }));
+    }
+
+    /// Flush updates all aggregated updates to the database.
+    pub fn flush(self, tx: &(impl DbTx + DbTxMut)) -> Result<(), reth_db::DatabaseError> {
+        if self.trie_operations.is_empty() {
+            return Ok(())
+        }
+
+        let mut account_trie_cursor = tx.cursor_write::<tables::AccountsTrie>()?;
+        let mut storage_trie_cursor = tx.cursor_dup_write::<tables::StoragesTrie>()?;
+
+        let mut trie_operations = Vec::from_iter(self.trie_operations);
+        trie_operations.sort_unstable_by(|a, b| a.0.cmp(&b.0));
+        for (key, operation) in trie_operations {
+            match key {
+                TrieKey::AccountNode(nibbles) => match operation {
+                    TrieOp::Delete => {
+                        if account_trie_cursor.seek_exact(nibbles)?.is_some() {
+                            account_trie_cursor.delete_current()?;
+                        }
+                    }
+                    TrieOp::Update(node) => {
+                        if !nibbles.0.is_empty() {
+                            account_trie_cursor.upsert(nibbles, StoredBranchNode(node))?;
+                        }
+                    }
+                },
+                TrieKey::StorageTrie(hashed_address) => match operation {
+                    TrieOp::Delete => {
+                        if storage_trie_cursor.seek_exact(hashed_address)?.is_some() {
+                            storage_trie_cursor.delete_current_duplicates()?;
+                        }
+                    }
+                    TrieOp::Update(..) => unreachable!("Cannot update full storage trie."),
+                },
+                TrieKey::StorageNode(hashed_address, nibbles) => {
+                    if !nibbles.is_empty() {
+                        // Delete the old entry if it exists.
+                        if storage_trie_cursor
+                            .seek_by_key_subkey(hashed_address, nibbles.clone())?
+                            .filter(|e| e.nibbles == nibbles)
+                            .is_some()
+                        {
+                            storage_trie_cursor.delete_current()?;
+                        }
+
+                        // The operation is an update, insert new entry.
+                        if let TrieOp::Update(node) = operation {
+                            storage_trie_cursor
+                                .upsert(hashed_address, StorageTrieEntry { nibbles, node })?;
+                        }
+                    }
+                }
+            };
+        }
+
+        Ok(())
+    }
+}
