+++ crates/trie/trie/src/node_iter.rs
@@ -1,5 +1,4 @@
 use crate::{hashed_cursor::HashedCursor, trie_cursor::TrieCursor, walker::TrieWalker};
-use reth_db::DatabaseError;
 use reth_primitives::{trie::Nibbles, B256};
 
 /// Represents a branch node in the trie.
@@ -84,7 +83,7 @@ where
     /// NOTE: The iteration will start from the key of the previous hashed entry if it was supplied.
     pub fn try_next(
         &mut self,
-    ) -> Result<Option<TrieElement<<H as HashedCursor>::Value>>, DatabaseError> {
+    ) -> Result<Option<TrieElement<<H as HashedCursor>::Value>>, C::Err> {
         loop {
             // If the walker has a key...
             if let Some(key) = self.walker.key() {
